Added an OperateOnMatrix utility class to work with matrixes. Moved utility classes to the outermost package Changed maps around
KNOWN ISSUES: Movement has some issues when dealing with the border conditions
TODO: change the controllers to be static nested classes, introduce JUnit and Javadocs for everything
 add a few design patterns
