+++ src/code/util/MatrixOperationListener.java
@@ -1,0 +1,6 @@
+package code.util;
+
+public interface MatrixOperationListener{
+    void doOperation(Object o, int i, int j);
+
+}
