+++ src/code/util/Pair.java
@@ -1,4 +1,4 @@
-package code.model.util;
+package code.util;
 
 public record Pair<X, Y>(X x, Y y) {
